create folder with components
