+++ frontend/packages/client/src/components/Community/CommunityCard/index.js
@@ -1,0 +1,1 @@
+export { default as CommunityCard } from './CommunityCard';
